custom hook for fetching games
