+++ game-hub/src/components/GameGrid.tsx
@@ -1,28 +1,11 @@
-import React, { useEffect } from "react";
-import { useState } from "react";
-import apiClient from "../services/api-client";
+
 import { Text } from "@chakra-ui/react";
+import useGames from "../hooks/useGames";
 
-interface Game {
-  id: number;
-  name: string;
-}
 
-interface FetchGameResponse {
-  count: number;
-  results: Game[];
-}
 
 const GameGrid = () => {
-  const [games, setGames] = useState<Game[]>([]);
-  const [error, setError] = useState("");
-
-  useEffect(() => {
-    apiClient
-      .get<FetchGameResponse>("x/games")
-      .then((res) => setGames(res.data.results))
-      .catch((err) => setError(err.message));
-  });
+  const {games, error} = useGames();
 
   return (
     <>
